Move the namespace algorithm functions to dedicated CacheNameSpaceTrait.
We intend to use the same namespacing algorithm for HPOS orders cache like we have been using before for posts. Moving this structure to a common trait allow us to use the same logic for both posts and HPOS.
