+++ plugins/woocommerce/includes/class-wc-order-factory.php
@@ -13,8 +13,6 @@ use Automattic\WooCommerce\Utilities\OrderUtil;
 
 defined( 'ABSPATH' ) || exit;
 
-use Automattic\WooCommerce\Utilities\OrderUtil;
-
 /**
  * Order factory class
  */
